refactor(declare): split out the display half, give local and readonly their own builtins
`declare.rs` held two unrelated jobs behind one options struct; the display
half moves to `declare/display.rs`. `local` and `readonly` get their own
command structs, so each accepts only the options it really takes. The
interpreter's two assignment targets are named rather than passed as three
loose booleans.
